backend/requirements.txt
asgiref==3.2.8
astroid==2.4.2
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.6.1
install==1.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
